Change default volume sizes created on first deploy (#3028)
Volumes of 1GB are fine when testing a new app and don't want to go over the free tier quote, but they are no useful when apps require larger machines that doesn't even fit within the free tier. This is more notorious with GPU apps where It is cumbersome to have to remember about `--volume-initial-size` when launching an app for the first time. 

This PR change the default volume size based on the machine size:
* `shared-cpu-1x` will continue to create 1GB volumes
*  `a100-40gb` and `a100-80gb` will default to 100GB
*  any other size will create 10GB volumes.

At the moment, after creating a 1GB volume,`fly deploy` recommends to use `fly volume extend` to increase it size, but it could be the case that the host where the volume landed doesn't have enough resources to do that on a second call and require migrating the volume somewhere else. By creating larger volume sizes it aims to reduce the risk of hitting a capacity issue early on the app onboarding.